* feature/merge-gods: (277 commits)
  Merge done, just add specific filed for visualization
  Added vendor
  Added strconv imports
  Add 1.13 and 1.14 versions to tests
  - fix arraylist test
  - fix utils test
  rename sort example to godsort
  README: fix BTree commit
  Update .travis.yml
  fix linkedhashmap.Map comment error
  typo fix
  utils: remove excessive type assertions
  Bump Travis versions
  TreeMap Floor and Ceiling functions
  - linked hashmap documentation and example
  - linked hash map core implementation + iterator + enumerable + tests + serialization (TODO: examples + documentation)
  - fix comments/documentation, rigorous serialization testing
  - linkedhashset simplification
  - update comments and documentation
  Sets bulk intialization
  ...

+++ utils/utils_test.go
@@ -1,6 +1,6 @@
-
-
-
+// Copyright (c) 2015, Emir Pasic. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
 
 package utils
 
@@ -74,7 +74,7 @@ func TestToStringFloats(t *testing.T) {
 	if actualValue, expectedValue := ToString(value), "1.123456"; !strings.HasPrefix(actualValue, expectedValue) {
 		t.Errorf("Got %v expected %v", actualValue, expectedValue)
 	}
-	value = float32(1.123456)
+	value = float64(1.123456)
 	if actualValue, expectedValue := ToString(value), "1.123456"; !strings.HasPrefix(actualValue, expectedValue) {
 		t.Errorf("Got %v expected %v", actualValue, expectedValue)
 	}
